backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.9.0
python-dotenv==1.0.0
openai==1.3.5
aiosqlite==0.19.0
python-multipart==0.0.6
yfinance>=0.2.33
